Split off tests for numpy functions.
