+++ astropy/utils/masked/tests/test_functions.py
@@ -1,0 +1,27 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+import numpy as np
+from numpy.testing import assert_array_equal
+
+
+from .test_masked import MaskedArraySetup, QuantitySetup, LongitudeSetup
+
+
+class TestMaskedArrayConcatenation(MaskedArraySetup):
+    def test_concatenate(self):
+        mb = self.mb[np.newaxis]
+        concat_a_b = np.concatenate((self.ma, mb), axis=0)
+        expected_data = np.concatenate((self.a, self.b[np.newaxis]), axis=0)
+        expected_mask = np.concatenate((self.mask_a, self.mask_b[np.newaxis]),
+                                       axis=0)
+        assert_array_equal(concat_a_b.unmasked, expected_data)
+        assert_array_equal(concat_a_b.mask, expected_mask)
+
+
+class TestMaskedQuantityConcatenation(TestMaskedArrayConcatenation,
+                                      QuantitySetup):
+    pass
+
+
+class TestMaskedLongitudeConcatenation(TestMaskedArrayConcatenation,
+                                       LongitudeSetup):
+    pass
