+++ astropy/utils/masked/tests/test_masked.py
@@ -555,24 +555,3 @@ class TestMaskedArrayRepr(MaskedArraySetup):
 
 class TestMaskedQuantityRepr(TestMaskedArrayRepr, QuantitySetup):
     pass
-
-
-class TestMaskedArrayConcatenation(MaskedArraySetup):
-    def test_concatenate(self):
-        mb = self.mb[np.newaxis]
-        concat_a_b = np.concatenate((self.ma, mb), axis=0)
-        expected_data = np.concatenate((self.a, self.b[np.newaxis]), axis=0)
-        expected_mask = np.concatenate((self.mask_a, self.mask_b[np.newaxis]),
-                                       axis=0)
-        assert_array_equal(concat_a_b.unmasked, expected_data)
-        assert_array_equal(concat_a_b.mask, expected_mask)
-
-
-class TestMaskedQuantityConcatenation(TestMaskedArrayConcatenation,
-                                      QuantitySetup):
-    pass
-
-
-class TestMaskedLongitudeConcatenation(TestMaskedArrayConcatenation,
-                                       LongitudeSetup):
-    pass
